requirements.txt
azure-core==1.35.0
azure-functions==1.23.0
azure-storage-blob==12.26.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
curl_cffi==0.12.0
frozendict==2.4.6
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
multitasking==0.0.12
numpy==2.2.6
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.65
